chore: Reorg core
